requirements.txt
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.68
langchain-groq==0.3.5
langchain-huggingface==0.3.0
langchain-text-splitters==0.3.8
langsmith==0.4.4
faiss-cpu==1.11.0
sentence-transformers==5.0.0
transformers==4.53.1
torch==2.7.1
huggingface-hub==0.33.2
tokenizers==0.21.2
safetensors==0.5.3
groq==0.29.0
pypdf==5.7.0
streamlit==1.46.1
python-dotenv==1.1.1
pydantic==2.11.7
numpy==2.3.1
requests==2.32.4
tqdm==4.67.1
setuptools==80.9.0